+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.2 - 2026-03-18
+### Added
+
+- Added `RandomNumberGenerator` class.
+
 ## 1.0.1 - 2026-03-09
 ### Added
 
@@ -14,9 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 1.0.0 - 2026-03-06
 ### Changed
 
-- Major readability improvements on whole code base
+- Major readability improvements on whole code base.
 
 ## 0.0.1 - 2022-08-24
 ### Added
 
-- Initial commit
+- Initial commit.